Extract oauth routes
